Refactored audio files
