+++ src/audio/mod.rs
@@ -1,130 +1,20 @@
+pub mod audio_unit_output;
+
 use crate::memory::memory::Memory;
 use crate::Word;
+use audio_unit_output::AudioUnitOutput;
 use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
 use cpal::{Device, Stream, SupportedStreamConfig};
 use std::sync::{Arc, RwLock};
 
-pub trait AudioUnit {
-    fn play_pulse(&mut self, stream_n: u8, frequency: f32, wave_duty_percent: f32);
-}
-
-pub struct CpalAudioUnit {
-    device: Device,
-
-    stream_1: Option<Stream>,
-    stream_2: Option<Stream>,
-    stream_3: Option<Stream>,
-    stream_4: Option<Stream>,
-}
-
-impl CpalAudioUnit {
-    pub fn new() -> Self {
-        let host = cpal::default_host();
-
-        let device = host
-            .default_output_device()
-            .expect("failed to find a default output device");
-
-        Self {
-            device,
-            stream_1: None,
-            stream_2: None,
-            stream_3: None,
-            stream_4: None,
-        }
-    }
-}
-
-impl AudioUnit for CpalAudioUnit {
-    fn play_pulse(&mut self, stream_n: u8, frequency: f32, wave_duty_percent: f32) {
-        let config = self.device.default_output_config().unwrap();
-
-        let stream = match config.sample_format() {
-            cpal::SampleFormat::F32 => {
-                run::<f32>(&self.device, &config.into(), frequency, wave_duty_percent).unwrap()
-            }
-            cpal::SampleFormat::I16 => {
-                run::<i16>(&self.device, &config.into(), frequency, wave_duty_percent).unwrap()
-            }
-            cpal::SampleFormat::U16 => {
-                run::<u16>(&self.device, &config.into(), frequency, wave_duty_percent).unwrap()
-            }
-        };
-
-        match stream_n {
-            1 => self.stream_1 = Some(stream),
-            2 => self.stream_2 = Some(stream),
-            _ => panic!("Non pulse stream given"),
-        }
-
-        fn run<T>(
-            device: &cpal::Device,
-            config: &cpal::StreamConfig,
-            frequency: f32,
-            wave_duty_percent: f32,
-        ) -> Result<Stream, anyhow::Error>
-        where
-            T: cpal::Sample,
-        {
-            let sample_rate = config.sample_rate.0 as f32;
-            let channels = config.channels as usize;
-
-            let mut sample_clock = 0f32;
-            let sample_in_period = sample_rate / frequency;
-            let high_part_max = sample_in_period * wave_duty_percent;
-
-            let mut next_value = move || {
-                sample_clock = (sample_clock + 1.0) % sample_rate; // 0..44099
-
-                if sample_clock % sample_in_period <= high_part_max {
-                    1.0
-                } else {
-                    -1.0
-                }
-            };
-
-            let err_fn = |err| println!("an error occurred on stream: {}", err);
-
-            let stream = device.build_output_stream(
-                config,
-                move |data: &mut [T], _: &cpal::OutputCallbackInfo| {
-                    for frame in data.chunks_mut(channels) {
-                        let next_value = next_value();
-                        let value: T = cpal::Sample::from::<f32>(&next_value);
-                        for sample in frame.iter_mut() {
-                            *sample = value;
-                        }
-                    }
-                },
-                err_fn,
-            )?;
-
-            Ok(stream)
-        }
-    }
-}
-
-pub struct OutputDebugAudioUnit {}
-
-impl AudioUnit for OutputDebugAudioUnit {
-    fn play_pulse(&mut self, stream_n: u8, frequency: f32, wave_duty_percent: f32) {
-        println!(
-            "S{}: Played at {} Hz, {}% duty",
-            stream_n,
-            frequency,
-            wave_duty_percent * 100.0
-        );
-    }
-}
-
-pub struct AudioUnitAdapter {
-    au: Box<dyn AudioUnit>,
+pub struct AudioUnit {
+    auo: Box<dyn AudioUnitOutput>,
     memory: Arc<RwLock<Memory>>,
 }
 
-impl AudioUnitAdapter {
-    pub fn new(au: Box<dyn AudioUnit>, memory: Arc<RwLock<Memory>>) -> Self {
-        Self { au, memory }
+impl AudioUnit {
+    pub fn new(au: Box<dyn AudioUnitOutput>, memory: Arc<RwLock<Memory>>) -> Self {
+        Self { auo: au, memory }
     }
 
     pub fn step(&mut self) {
@@ -141,7 +31,9 @@ impl AudioUnitAdapter {
         let all_sound_trigger = nr52 & 0b10000000 == 0b10000000;
 
         if !all_sound_trigger {
-            // FIXME: Maybe we need to stop sounds
+            self.auo.stop_all();
+
+            // FIXME: Might need to restart control registers
             return;
         }
 
@@ -201,6 +93,6 @@ impl AudioUnitAdapter {
             _ => panic!("Invalid Wave Duty"),
         };
 
-        self.au.play_pulse(wave_n, frequency, wave_duty_percent);
+        self.auo.play_pulse(wave_n, frequency, wave_duty_percent);
     }
 }
